Prod setup refactoring (#15)
* Serve from index file server directory

* Refactor server code

* Server scripts for deployment

* Deploy
