+++ package.json
@@ -1,15 +1,15 @@
 {
   "name": "mern-kanban",
-  "version": "1.0.0",
-  "description": "A full stack application with React and MongoDB",
-  "main": "index.js",
+  "version": "2.0.0",
+  "description": "A full stack application built with React, Express, Node.js and MongoDB",
+  "main": "index.jsx",
   "scripts": {
     "build": "webpack -p",
     "deploy": "npm run build",
     "dev": "webpack-dev-server --open",
     "postinstall": "npm install --save-dev webpack-cli",
-    "server": "node src/server/server.mjs",
-    "server-dev": "nodemon src/server/server --exec babel-node src/server/server",
+    "server": "node src/server/index.mjs",
+    "server-dev": "nodemon src/server/index.mjs --exec babel-node src/server/index.mjs",
     "start-dev": "concurrently \"npm run server-dev\" \"npm run dev\"",
     "server-spec": "babel-node src/server/server.spec",
     "start": "npm run server"
